move misc stories under src directory
