+++ src/FilterAdder.stories.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import { applyDefaults } from './utils/schema'
 import { useTheme } from './utils/theme'
-import ThemePicker from '../stories/themePicker'
+import ThemePicker from './stories/themePicker'
 import { FilterAdder } from '.'
 
 let mockTree = {
