+++ src/exampleTypes/Geo.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import AsyncSelect from 'react-select/lib/Async'
 import { loadHereOptions, geoCodeLocation } from '../utils/geo'
-import ThemePicker from '../../stories/themePicker'
+import ThemePicker from '../stories/themePicker'
 import { Flex } from '../greyVest'
 import TestTree from './stories/testTree'
 import { Geo } from '.'
