+++ src/exampleTypes/Number.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import TestTree from './stories/testTree'
-import ThemePicker from '../../stories/themePicker'
+import ThemePicker from '../stories/themePicker'
 import { Number } from '.'
 
 storiesOf('Components|ExampleTypes', module)
