+++ src/purgatory/CheckButton.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { CheckButton } from '.'
-import ThemePicker from '../../stories/themePicker'
+import ThemePicker from '../stories/themePicker'
 
 storiesOf('Components|Search Components/Internals/Checkbutton', module)
   .addDecorator(ThemePicker('greyVest'))
