+++ src/purgatory/ToggleFiltersButton.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import ThemePicker from '../../stories/themePicker'
+import ThemePicker from '../stories/themePicker'
 import { ToggleFiltersButton } from '.'
 
 let click = action('clicked')
