+++ src/purgatory/TreePauseButton.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import { observable } from 'mobx'
-import ThemePicker from '../../stories/themePicker'
+import ThemePicker from '../stories/themePicker'
 import { TreePauseButton } from '.'
 import { SearchTree } from '..'
