+++ src/queryBuilder/QueryBuilder.stories.js
@@ -4,7 +4,7 @@ import { exampleTypes, mockService } from 'contexture-client'
 import ContextureMobx from '../utils/contexture-mobx'
 import { TypeMap } from '../exampleTypes'
 import { componentForType } from '../utils/schema'
-import ThemePicker from '../../stories/themePicker'
+import ThemePicker from '../stories/themePicker'
 import QueryBuilder from '.'
 
 let Client = ContextureMobx({
